+++ include/ozz/animation/offline/raw_track.h
@@ -60,8 +60,30 @@ struct RawTrackKeyframe {
 
 namespace internal {
 
-// Offline user-channel animation track type internal implementation. See
-// Raw*Track for more details.
+// Offline user-channel animation track type implementation.
+// This offline track data structure is meant to be used for user-channel
+// tracks, aka animation of variables that aren't joint transformation. It is
+// available for tracks of 1 to 4 floats (RawFloatTrack, RawFloat2Track, ...,
+// RawFloat4Track) and quaternions (RawQuaternionTrack). Quaternions differ from
+// float4 because of the specific interpolation and comparison treatment they
+// require. As all other Raw data types, they are not intended to be used in run
+// time. They are used to define the offline track object that can be converted
+// to the runtime one using the a ozz::animation::offline::TrackBuilder. This
+// animation structure exposes a single sequence of keyframes. Keyframes are
+// defined with a ratio, a value and an interpolation mode:
+// - Ratio: A track has no duration, so it uses ratios between 0 (beginning of
+// the track) and 1 (the end), instead of times. This allows to avoid any
+// discrepancy between the durations of tracks and the animation they match
+// with.
+// - Value: The animated value (float, ... float4, quaternion).
+// - Interpolation mode (`ozz::animation::offline::RawTrackInterpolation`):
+// Defines how value is interpolated with the next key. Track structure is then
+// a sorted vector of keyframes. RawTrack structure exposes a Validate()
+// function to check that all the following rules are respected:
+// 1. Keyframes' ratios are sorted in a strict ascending order.
+// 2. Keyframes' ratios are all within [0,1] range.
+// RawTrack that would fail this validation will fail to be converted by
+// the RawTrackBuilder.
 template <typename _ValueType>
 struct RawTrack {
   typedef _ValueType ValueType;
@@ -92,25 +114,7 @@ struct RawTrack {
 };
 }  // namespace internal
 
-// Offline user-channel animation track type implementation.
-// This data type is meant to be used for user-channel tracks, aka animation of
-// variables that aren't joint transformation. It is available for tracks of 1
-// to 4 floats (RawFloatTrack, RawFloat2Track, ..., RawFloat4Track) and
-// quaternions (RawQuaternionTrack). Quaternions differ from float4 because of
-// the specific interpolation and comparison treatment they require.
-// As all other Raw data types, they are not intended to be used in run time.
-// They are used to define the offline track object that can be converted to the
-// runtime one using the a ozz::animation::offline::TrackBuilder.
-// A track has no duration. It uses ratios between 0 (beginning of the track)
-// and 1 (the end), instead of times. This allows to avoid any discrepancy
-// between the durations of tracks and the animation they match with.
-// This animation structure exposes a single sequence of keyframes. Keyframes
-// are defined with a ratio, a value and an interpolation mode. Track structure
-// is then a sorted vector of keyframes.
-//  1. Keyframes' ratios are sorted in a strict ascending order.
-//  2. Keyframes' ratios are all within [0,1] range.
-// RawTrack that would fail this validation will fail to be converted by
-// the RawTrackBuilder.
+// Offline user-channel animation track type instantiation.
 struct RawFloatTrack : public internal::RawTrack<float> {};
 struct RawFloat2Track : public internal::RawTrack<math::Float2> {};
 struct RawFloat3Track : public internal::RawTrack<math::Float3> {};