+++ samples/framework/utils.h
@@ -51,24 +51,25 @@ struct Mesh;
 
 // Utility class that helps with controlling animation playback time. Time is
 // computed every update according to the dt given by the caller, playback speed
-// and "play" state. OnGui function allows to tweak controller parameters
-// through the application Gui.
+// and "play" state.
+// Internally time is stored as a ratio in unit interval [0,1], as expected by
+// ozz runtime animation jobs.
+// OnGui function allows to tweak controller parameters through the application
+// Gui.
 class PlaybackController {
  public:
   // Constructor.
   PlaybackController();
 
   // Sets animation current time.
-  void set_time(float _time) {
-    previous_time_ = time_;
-    time_ = _time;
-  }
+  void set_time_ratio(float _time);
 
   // Gets animation current time.
-  float time() const { return time_; }
+  float time_ratio() const;
 
-  // Gets animation time of last update.
-  float previous_time() const { return previous_time_; }
+  // Gets animation time ratio of last update. Useful when the range between
+  // previous and current frame needs to pe processed.
+  float previous_time_ratio() const;
 
   // Sets playback speed.
   void set_playback_speed(float _speed) { playback_speed_ = _speed; }
@@ -96,11 +97,12 @@ class PlaybackController {
              bool _enabled = true, bool _allow_set_time = true);
 
  private:
-  // Time of the previous update.
-  float previous_time_;
+  // Current animation time ratio, in the unit interval [0,1], where 0 is the
+  // beginning of the animation, 1 is the end.
+  float time_ratio_;
 
-  // Current animation time.
-  float time_;
+  // Time ratio of the previous update.
+  float previous_time_ratio_;
 
   // Playback speed, can be negative in order to play the animation backward.
   float playback_speed_;